pyproject.toml
[tool.poetry]
name = "hack-dbeaver-credentials"
version = "0.1.0"
description = "Set DBEaver credentials in a simple command"
authors = ["Mateus Andrade <msn.mateus@hotmail.com>"]
readme = "README.md"
packages = [{include = "hack_dbeaver_credentials"}]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.7.0"}
pycryptodome = "^3.16.0"
prettyconf = "^2.2.1"
rich = "^12.6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.11.4"
black = "^22.12.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
